refactor: vercel에 맞게 UI 수정
