+++ src/components/BottomSheet/BottomSheet.module.scss
@@ -12,7 +12,7 @@
 .sheet {
   position: fixed;
   left: 50%;
-  bottom: 0;
+  bottom: 186px;
   transform: translateX(-50%);
   width: 375px;
   max-width: 100%;
@@ -50,6 +50,13 @@
   font-size: 20px;
   line-height: 150%;
   cursor: pointer;
+}
+
+.close img {
+  width: 24px;
+  height: 24px;
+  display: block;
+  object-fit: contain;
 }
 .content {
   padding: 12px 20px 0;
