+++ src/components/BottomSheet/CategorySheet.module.scss
@@ -25,7 +25,7 @@
   position: fixed;
   left: 0;
   right: 0;
-  bottom: 0;
+  bottom: 186px;
   margin: 0 auto;
   transform: none;
   width: var(--FRAME-W, 375px);
@@ -73,6 +73,7 @@
   width: 24px;
   height: 24px;
   display: block;
+  object-fit: contain;
 }
 /* grid */
 .grid {
