+++ src/components/TopBar.module.scss
@@ -7,7 +7,7 @@
 }
 
 .root {
-  position: fixed;
+  position: absolute;
   top: 0;
   left: 50%;
   transform: translateX(-50%);
