+++ src/layouts/OnboardingLayout.module.scss
@@ -7,8 +7,8 @@
 .iphoneFrame {
   position: relative;
   width: 100%;
-  height: 100vh;
   max-width: 375px;
+  height: 812px;
   margin: 0 auto;
   background: #fff;
   overflow: hidden;
