+++ src/main.tsx
@@ -2,6 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
 
+import "./index.css";
 import "./styles/main.scss";
 
 createRoot(document.getElementById("root")!).render(
