+++ src/pages/Loading/LoadingPage.module.scss
@@ -22,7 +22,7 @@
 
 .animationContainer {
   position: absolute;
-  bottom: 0;
+  bottom: -4px;
 }
 
 .textContainer {
