+++ src/pages/SelectPlace/SelectPlace.module.scss
@@ -1,28 +1,22 @@
 .selectPlace {
   width: 100%;
-  min-height: 100vh;
   display: flex;
   flex-direction: column;
   background: #fff;
-  margin-top: 40px;
   position: relative;
   overflow: hidden;
 }
 
 .header {
-  position: fixed;
-  top: 100px;
-  left: 20px;
-  right: 20px;
   z-index: 10000;
-  background: #fff;
+  margin-bottom: 16px;
 }
 
 .title {
   font-size: 18px;
   font-weight: 600;
   color: #383325;
-  margin: 0 0 8px 0;
+  margin: 0 0 4px 0;
   line-height: 150%;
   white-space: pre-line;
 }
@@ -38,10 +32,6 @@
 }
 
 .mapContainer {
-  position: fixed;
-  top: 217px;
-  left: 20px;
-  right: 20px;
   height: 240px;
   border-radius: 8px;
   overflow: hidden;
@@ -56,12 +46,9 @@
 }
 
 .alertSection {
-  position: fixed;
-  top: 465px; /* 지도 아래 (217px + 240px + 20px 간격) */
-  left: 20px;
-  right: 20px;
   display: flex;
   gap: 2px;
+  margin-top: 6px;
   z-index: 10000;
 }
 
@@ -76,10 +63,11 @@
 
 .mapExpanded {
   position: fixed;
-  top: 60px; /* TopBar 높이만큼 아래에서 시작 */
-  left: 0;
-  width: 100vw;
-  height: calc(100vh - 60px); /* TopBar 높이만큼 제외 */
+  top: 52px; /* TopBar 높이만큼 아래에서 시작 */
+  left: 50%;
+  transform: translateX(-50%);
+  width: 375px; /* 375 * 812 화면에 맞춤 */
+  height: 752px; /* 812px - 60px(TopBar) = 752px */
   z-index: 9999;
   background: #fff;
   margin: 0;
@@ -88,8 +76,8 @@
 }
 
 .mapExpanded .map {
-  width: 100vw;
-  height: calc(100vh - 60px);
+  width: 375px;
+  height: 752px;
   margin: 0;
 }
 
@@ -307,10 +295,6 @@
 }
 
 .expandedAddedLocationCard {
-  position: fixed;
-  bottom: 200px; /* 하얀 창 위 - 동적으로 조정됨 */
-  left: 50%;
-  transform: translateX(-50%);
   display: flex;
   width: 335px;
   height: 54px;
@@ -322,7 +306,7 @@
   background: #ffd4ca;
   box-shadow: 0 4px 30px 0 rgba(0, 48, 120, 0.1);
   z-index: 10001; /* 지도보다 위에 표시 */
-  margin-bottom: 5px; /* 하얀 창과의 간격 */
+  margin-top: 390px;
 }
 
 .addedLocationCard span,
@@ -351,20 +335,29 @@
 
 .nextButton {
   position: fixed;
-  bottom: 44px;
-  left: 20px;
-  right: 20px;
-  padding: 16px;
-  background: #ccc;
+  bottom: 230px;
+  left: 50%;
+  transform: translateX(-50%);
+  display: flex;
+  width: 335px;
+  height: 54px;
+  padding: 10px;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  flex-shrink: 0;
+  border-radius: 10px;
+  background: #d9d9d9;
   border: none;
-  border-radius: 8px;
-  color: #666;
+  color: #b5b5b5;
   font-size: 16px;
   font-weight: 600;
   line-height: 150%;
   cursor: not-allowed;
   transition: all 0.2s;
   z-index: 10000;
+  /* ButtonEffect */
+  box-shadow: 0 4px 30px 0 rgba(120, 0, 0, 0.1);
 }
 
 .nextButton.active {
